// src/app/model/register.ts
export interface Register {
  parentUserId: number;
  studentUserId: number;
  addressId: number;
  firstName: string;
  lastName: string;
  dateOfBirth: string;
  gender: string;
  classId: number;
  motherName: string;
  motherPhoneNumber: string;
  motherOccupation: string;
  fatherName: string;
  fatherPhoneNumber: string;
  fatherOccupation: string;
  email: string;
  doorNum: string;
  street: string;
  addrLine: string;
  city: string;
  state: string;
  pincode: string;
  studentUsername: string;
  studentPassword: string;
  parentUsername: string;
  parentPassword: string;
}